+++ projects/_03_independent_cqc/unittest_data/ind_cqc_test_file_schemas.py
@@ -3144,19 +3144,6 @@ class IndCQCDataUtils:
         ]
     )
 
-    allocate_primary_service_type_second_level_schema = StructType(
-        [
-            StructField(CQCL.location_id, StringType(), True),
-            StructField(CQCLClean.services_offered, ArrayType(StringType(), True)),
-        ]
-    )
-    expected_allocate_primary_service_type_second_level_schema = StructType(
-        [
-            *allocate_primary_service_type_second_level_schema,
-            StructField(IndCQC.primary_service_type_second_level, StringType(), True),
-        ]
-    )
-
 
 @dataclass
 class NullCtPostsToBedsOutliers: